+++ db/sql/9665__drop_create_table_anteroapi_opiskelijat_ja_tutkinnot.sql
@@ -22,7 +22,7 @@ CREATE TABLE [api].[opiskelijat_ja_tutkinnot](
 	[Sukupuoli] [nvarchar](100) NULL,
 	[Tutkinnot] [int] NULL,
 	[Opiskelijat] [int] NULL,
-	[opiskelijatFte] [int] NULL,
+	[opiskelijatFte] [float] NULL,
 	[Uudet opiskelijat] [int] NULL,
 	[Koulutusaste taso 1] [nvarchar](200) NULL,
 	[Koulutusaste taso 2] [nvarchar](200) NULL,
